解除循环依赖
