+++ packages/app-backend/src/router.js
@@ -75,19 +75,3 @@ const clipRoute = function (route) {
     }),
   }
 }
-
-export function getCustomRouterDetails(router) {
-  return {
-    _custom: {
-      type: 'router',
-      display: 'VueRouter',
-      value: {
-        options: router.options,
-        currentRoute: router.currentRoute,
-      },
-      fields: {
-        abstract: true,
-      },
-    },
-  }
-}
