+++ packages/shared-utils/src/index.js
@@ -1,1 +1,2 @@
 export * from './api'
+export { getCatchedGetters } from './details/custom'
