+++ packages/shared-utils/src/util.js
@@ -3,8 +3,7 @@ import { isRef } from 'vue'
 import { transform, cloneDeepWith, cloneDeep } from 'lodash-es'
 import * as CircularJSON from './transfer'
 import { getCustomInstanceDetails } from '@back/process'
-import { getCustomStoreDetails } from '@back/vuex'
-import { getCustomRouterDetails } from '@back/router'
+import { getCustomRouterDetails, getCustomStoreDetails } from './details/custom'
 import SharedData from './shared-data'
 import { isChrome } from './env'
